+++ src/detections/configs.rs
@@ -77,6 +77,8 @@ pub struct StoredStatic {
     pub output_path: Option<PathBuf>,
     pub common_options: CommonOptions,
     pub multiline_flag: bool,
+    pub include_computer: HashSet<CompactString>,
+    pub exclude_computer: HashSet<CompactString>,
 }
 impl StoredStatic {
     /// main.rsでパースした情報からデータを格納する関数
@@ -322,6 +324,98 @@ impl StoredStatic {
         } else {
             TargetIds::default()
         };
+        let include_computer: HashSet<CompactString> = match &input_config.as_ref().unwrap().action
+        {
+            Some(Action::CsvTimeline(opt)) => opt
+                .output_options
+                .detect_common_options
+                .include_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::JsonTimeline(opt)) => opt
+                .output_options
+                .detect_common_options
+                .include_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::Metrics(opt)) => opt
+                .detect_common_options
+                .include_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::PivotKeywordsList(opt)) => opt
+                .detect_common_options
+                .include_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::LogonSummary(opt)) => opt
+                .detect_common_options
+                .include_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            _ => HashSet::default(),
+        };
+        let exclude_computer: HashSet<CompactString> = match &input_config.as_ref().unwrap().action
+        {
+            Some(Action::CsvTimeline(opt)) => opt
+                .output_options
+                .detect_common_options
+                .exclude_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::JsonTimeline(opt)) => opt
+                .output_options
+                .detect_common_options
+                .exclude_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::Metrics(opt)) => opt
+                .detect_common_options
+                .exclude_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::PivotKeywordsList(opt)) => opt
+                .detect_common_options
+                .exclude_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            Some(Action::LogonSummary(opt)) => opt
+                .detect_common_options
+                .exclude_computer
+                .as_ref()
+                .unwrap_or(&vec![])
+                .iter()
+                .map(CompactString::from)
+                .collect(),
+            _ => HashSet::default(),
+        };
         let mut ret = StoredStatic {
             config: input_config.as_ref().unwrap().to_owned(),
             config_path: config_path.to_path_buf(),
@@ -428,6 +522,8 @@ impl StoredStatic {
             output_path: output_path.cloned(),
             common_options,
             multiline_flag,
+            include_computer,
+            exclude_computer,
         };
         ret.profiles = load_profile(
             check_setting_path(
@@ -713,6 +809,14 @@ pub struct DetectCommonOption {
     /// Output verbose information
     #[arg(help_heading = Some("Display Settings"), short = 'v', long, display_order = 480)]
     pub verbose: bool,
+
+    /// Scan only these computer names (ex: ComputerA) (ex: ComputerA,ComputerB)
+    #[arg(help_heading = Some("Filtering"), long = "include-computer", value_name = "COMPUTER", conflicts_with = "exclude-computer", use_value_delimiter = true, value_delimiter = ',', display_order = 351)]
+    pub include_computer: Option<Vec<String>>,
+
+    /// Do not scan these computer names (ex: ComputerA) (ex: ComputerA,ComputerB)
+    #[arg(help_heading = Some("Filtering"), long = "exclude-computer", value_name = "COMPUTER", conflicts_with = "include_computer",use_value_delimiter = true, value_delimiter = ',', display_order = 314)]
+    pub exclude_computer: Option<Vec<String>>,
 }
 
 #[derive(Args, Clone, Debug)]
@@ -1758,6 +1862,8 @@ fn extract_output_options(config: &Config) -> Option<OutputOption> {
                 quiet_errors: option.quiet_errors,
                 config: option.config.clone(),
                 verbose: option.verbose,
+                include_computer: None,
+                exclude_computer: None,
             },
             exact_level: None,
             enable_unsupported_rules: false,
@@ -1802,6 +1908,8 @@ fn extract_output_options(config: &Config) -> Option<OutputOption> {
                 config: Path::new("./rules/config").to_path_buf(),
                 verbose: false,
                 json_input: false,
+                include_computer: None,
+                exclude_computer: None,
             },
             enable_unsupported_rules: false,
             clobber: false,
@@ -1845,6 +1953,8 @@ fn extract_output_options(config: &Config) -> Option<OutputOption> {
                 config: Path::new("./rules/config").to_path_buf(),
                 verbose: false,
                 json_input: false,
+                include_computer: None,
+                exclude_computer: None,
             },
             enable_unsupported_rules: true,
             clobber: false,